rename: search 관련 api, type 파일 위치 변경
