+++ src/entities/search/index.ts
@@ -1,0 +1,3 @@
+export { getRecentSearchList } from './api/search';
+
+export type { RecentSearchRegion, SearchPageState } from './model/types';
