+++ src/entities/search/model/types.ts
@@ -1,15 +1,12 @@
-export type RecentSearchListDto = Array<{
+export type RecentSearchRegion = {
   city: string;
   district: string;
-}>;
+};
+
+export type RecentSearchListDto = Array<RecentSearchRegion>;
 
 export type SearchPageState =
   | {
       mode?: 'memberSetting';
     }
   | undefined;
-
-export type RecentSearchRegion = {
-  city: string;
-  district: string;
-};
