+++ src/pages/User/TabSection/SettingList/RegionSetMenu/RegionSetMenu.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 
-import { SearchPageState } from '@/features/search';
+import { SearchPageState } from '@/entities/search';
 
 import { useAppSelector } from '@/shared/lib/useAppSelector';
 import { LocationIcon } from '@/shared/ui';
